+++ src/socket.h
@@ -5,6 +5,7 @@
 
 
 #include <stdbool.h>
+#include <stdint.h>
 
 
 
@@ -14,10 +15,10 @@
 
 
 
-bool    cwdaemon_initialize_socket(cwdaemon_t * cwdaemon, int network_port);
+bool    cwdaemon_initialize_socket(cwdaemon_t * cwdaemon, uint16_t network_port);
 void    cwdaemon_close_socket(cwdaemon_t * cwdaemon);
 ssize_t cwdaemon_sendto(cwdaemon_t * cwdaemon, const char * reply);
-int     cwdaemon_recvfrom(cwdaemon_t * cwdaemon, char * request, int size);
+ssize_t cwdaemon_recvfrom(cwdaemon_t * cwdaemon, char * request, int size);
